modified ChanceHelper approach and added IChanceObjectD for doubles, IChanceObject is now ints

+++ CommonUtils/CommonUtils.Core/Interfaces/IChanceObject.cs
@@ -2,5 +2,5 @@
 
 public interface IChanceObject
 {
-    public double Chance { get; set; }
+    public int Chance { get; set; }
 }
